+++ resources/views/modules/hr/payroll.blade.php
@@ -1,8 +1,3 @@
-@extends('layouts.backoffice')
-
-@section('title', 'HR & Payroll')
-
-@section('content')
 <div class="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-6">
     {{-- Header --}}
     <div class="flex flex-col sm:flex-row sm:items-center sm:justify-between mb-6">
@@ -112,4 +107,3 @@
         </table>
     </div>
 </div>
-@endsection